seeded

+++ backend/db/migrations/20230514155014-create-spot.js
@@ -37,11 +37,11 @@ module.exports = {
         allowNull: false
       },
       lat: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.DECIMAL(8,6),
         allowNull: false
       },
       lng: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.DECIMAL (9,6),
         allowNull: false,
       },
       name: {
